initPopup placed in a separate module
